requirements.txt
yfinance
pandas
requests
python-dotenv
streamlit
plotly
numpy
